[Property delegates] Allow the backing property access to be specified.
The backing property for a property with a delegate will have an
access level that is, by default, the minimum of `internal` and the
access level of the original property. The access level of the setter
is computed similarly, but can be affected by (e.g.) private(set).

The default access level of the backing property can be adjusted by
placing an access specifier after the "by" keyword, e.g., "by private
Foo", so long as this access level does not exceed that of the
original property.