[closure-lifetime] Add a fix_lifetime to communicate the end of lifetime of the closure.
Otherwise, we have ownership expressed via a control flow dependence instead of
a data dependence. This causes from an ownership issues since from a semantics
perspective, the trivial closure we use is not related to the parent operand. So
a pass like Semantic ARC Opts that uses def-use edges will remove such copies as
being unneeded. From an ownership semantics perspective, the fix_lifetime acts
as an artificial "regular use".