multifile: make protocol-conformance-member pass on Windows
PE/COFF does not include a symbol table in the generated executable
binary.  Instead, use `public` to expose the getter, and then use
`-coff-exports` from `llvm-readobj` to get the symbols that are
exported.  Fortunately, the same tool can be used to list the symbol
table contents for ELF and MachO binary.  This allows us to share the
test across all the targets.