DebugInfo: deal with absolute path on Windows
Windows absolute paths are either UNC paths (begin with \\) or are
rooted with a drive letter.  Unfortunately, we cannot conditionalise the
path itself on the OS type.  On Unices // indicates the alternate root
path as per POSIX.  However, Linux does not implement this and // is
treated as /.  Use this to create a path that appears to be absolute on
all the targets.