[android] Fix IRGen/condfail.sil test for Android ARMv7/AArch64.
For AArch64, the fix is duplicating the arm64 lines, since those are not
taken when testing Android/Linux which use AArch64 as the architecture.

For ARMv7, the change is supporting both trap and .inst 0xe7ffdefe.
According to llvm/lib/Target/ARM/ARMAsmPrinter.cpp, non-Darwin binutils
do not support the mnemonic trap, so an .inst is emitted instead. The
same instruction has to be used in both places, though.

For both architectures add the Android version of APP/NO_APP, which uses
@-symbols or // instead of ##.