build: attempt to dynamically detect libdispatch
We know during the build whether we are building libdispatch for the
toolchain or not.  This wires that into the test suite.  From a quick
inspection, it seems that the tests currently do not depend on the Swift
SDK overlay for libdispatch.  We can easily wire the library path for
libdispatch to the tests through the substitutions that we perform
through lit.