[Property delegates] Model delegate backing storage access as a “target”
Rather than hardcoding the access patterns for the backing property of
a property with a delegate, extend the “target implementation” notion
used to abstract property access for code synthesis to understand the
delegate pattern. This collapses the getter/setter implementation into
the “trivial” cases uses the delegate access pattern.

As a bonus, we get @_transparent when it makes sense.