[android] Fix PIC test for Android ARMv7/AArch64
For the AArch64 test, the only change is making the @PAGE piece
optional. LLVM doesn't seem to generate that part when targetting
Android or Linux.

For the ARMv7 tests, it is a little bit more complicated. Android seems
to use a different structure for PIC, which involve two labels and
instructions which are not movw/movt. To keep the iOS test undisturbed,
I added an extra check for the target-sdk-name. That allows iOS to be
the same, and Android to have their own checks. For Linux (the only
other SDK I can find that targets ARMv7), I used the Android solution,
because that seems to be what LLVM generates when using
armv7-none-linux.