[update-checkout] Don't use env for portability.
env is not portable in systems that do not have env, like for example Windows. However, passing the environment is supported by shell.run using the env parameter, which should be portable.

For some reason shell.run doesn't merge the environment with the current environment by default, so we do it externally to not modify other possible usages of the function.