[ownership] Fix a corner case in the linear lifetime checker.
Specifically, if we had a value that was consumed in the same block that it was
produced in, we ignored use-after-frees in subsequent blocks. We did check for
use-after-frees in the same block though. The general case was handled
correctly, this was just an incorrect early exit.

On master, this only found one problem (namely the one fixed in:
14d39c0cf951aaca3b8d782b0738bf55bf6672ab). I do not expect this corner case to
have more impact after cherry-picking to 5.1.

rdar://49794321
(cherry picked from commit 6552a0221ef2514b0f9fe3d79a314ffc02870335)