[Python][TF-336] Extends unit testing infra to support several python versions
In order to run the tests using different Python versions, two new
lit-directives were added: target-run-python2-interop-simple-swift and
target-run-python3-interop-simple-swift. Each of them behave exactly
like target-run-simple-swift only that they set the env variable
PYTHON_VERSION with the desired major python version to be loaded.

As well, within `python_runtime.swift` there is logic to check
if a given PYTHON_VERSION is loadable. All tests in the file
are short-circuited if the desired python library wasn't. Downside
is that this not play nice with lit failing error's reporting.

In this solution there is no duplication of logic in `lit.cfg`
to search for paths to different versions of the Python library
files as it only provides directives that set `PYTHON_VERSION`
before running tests.