[ParseableInterfaces] Always load valid modules from the PIML
Previously, we would return `errc::not_supported` when there was a
valid .swiftmodule that we could load instead of compiling from an
interface. This worked great, when there was only one place in the
search paths where there's a .swiftmodule or .swiftinterface we could
load. We'd skip all the valid ones and then let the
SerializedModuleLoader load them instead.

However, when there's a valid .swiftinterface anywhere in the search
paths, suddenly we'll load from that instead of the valid .swiftmodule
we saw previously. Instead of returning `not_supported` and possibly
searching the rest of the paths for something that might not be right
for a local build, just pass up the module we've already loaded,
validated, and made sure is up-to-date.

Also add some more logging throughout `discoverLoadableModule` so we can
more easily catch issues in the future.

Fixes rdar://49479386