IRGen: Fix missing nominal type descriptor when emitting lazy conformances
Conformance descriptors are emitted after the lazy emission queues are
drained, so need to queue up anything referenced from a conformance
descriptor before we queue up the conformance descriptor itself.

The existing logic was only forcing foreign type metadata, which is
wrong. We only ever need the nominal type descriptor and not the
metadata; and we need it for all types.

Note that a better fix for this specific test case would be to not
emit the conformance at all; if the conforming type or the protocol
is not referenced there's no way to reference the conformance either.
But that would be a bigger change.

Fixes <rdar://problem/49710077>.