[build-script] Rename is_build_script_impl_product to needs_toolchain.
While the previous name is factual at the moment, I think needs_toolchain
describes better why those products are special.

Sadly Ninja has to be always a special product, and it doesn't fit
needs_toolchain, but it also doesn't need to be handled with the other
products, so answering False for it is fine.

The rest of the changes are mechanical, and flipping the logic around.