[build-script] Extract most parts related to build-script-impl
Most parts related to build-script-impl moved into its own type
BuildScriptImplHelper. This will make moving this helper to its own file
and to a builder easier later.

The parts that haven't moved are the path of the script (still a
constant in the main build-script file), and the invocation to check the
migration parameters. Those will be dealt later.

The changes move the methods that turn build-script args into
build-script-impl args and creates the environment variables into this
new helper. It also moves the specific of building action names and how
to invoke the script from their previous places in BuildScriptInvocation
into the Helper.